cover_letter_generator/cover_letter_generator/requirements.txt
streamlit
python-dotenv
PyPDF2
openai
langchain
langchain_groq
